services/motechrunner/requirements.txt
Django==1.10.5
requests==2.12.5
jsonobject==0.7.1
psycopg2==2.6.2
PyYAML==3.12
django-jsonfield==1.0.1
pytz==2016.10
